requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
pydantic>=1.10.0
pyyaml>=6.0

# Data processing
pyarrow>=10.0.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# ML/AI
scikit-learn>=1.1.0
openai>=1.0.0
langchain>=0.1.0
chromadb>=0.4.0

# Workflow orchestration
apache-airflow>=2.5.0

# Utilities
click>=8.0.0
tqdm>=4.64.0
